created separate transfer utility file
