+++ lib/utils/custom_transfers/common.dart
@@ -1,0 +1,3 @@
+List<List<int>> boolToIntBitmap(List<List<bool>> bitmap) {
+  return bitmap.map((row) => row.map((b) => b ? 1 : 0).toList()).toList();
+}
